# [ONEDNN] fix int8 resnet50 accuracy by disabling inplace