@xinyu-intel 2023-04-10
questions:

1) will post-binary bring performance regression compared with post-sum?
2) does paddle provide utils for checking if a tensor can be inplaced?