@YangQun1 2023-04-11
> 1. will post-binary bring performance regression compared with post-sum?

I tested rn50 and mobilenetv2 locally, and found that f32 model has performance regression, while int8 model has performance improvement. May need validation team help to verify more models.

> 2. does paddle provide utils for checking if a tensor can be inplaced?

I didn't find such utils. will check it further